Refactor item controller to use item interface
- Refactor item controller index and optionsIndex methods to use item interface.
- Moved around a few config files.
- Localisation missing for item name.
