+++ resources/lang/en/item-type-allocated-expense/allowed-values.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'description-prefix-year' => 'Include results with effective date in ',
+
+    'description-prefix-month' => 'Include results with effective date in ',
+];
