+++ resources/lang/en/item-type-allocated-expense/parameters.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'title-include-unpublished' => 'Include unpublished expenses?',
+    'description-include-unpublished' => 'Optionally include any unpublished expenses',
+
+    'title-year' => 'Year',
+    'description-year' => 'Show results for the requested year',
+
+    'title-month' => 'Month',
+    'description-month' => 'Show results fo the requested month'
+];
