+++ resources/lang/en/item/allowed-values.php
@@ -3,10 +3,6 @@
 declare(strict_types=1);
 
 return [
-    'description-prefix-year' => 'Include results with effective date in ',
-
-    'description-prefix-month' => 'Include results with effective date in ',
-
     'description-prefix-category' => 'Include results with items assigned to \'',
     'description-suffix-category' => '\' category',
