--- resources/lang/en/item/fields.php
@@ -1,20 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-return [
-    'title-description' => 'Item description',
-    'description-description' => 'Enter a description for the item',
-
-    'title-effective_date' => 'Item effective date',
-    'description-effective_date' => 'Enter an effective date for the item',
-
-    'title-publish_after' => 'Publish after this date',
-    'description-publish_after' => 'Enter a date to withhold publishing until',
-
-    'title-total' => 'Item total',
-    'description-total' => 'Enter the total cost for the item',
-
-    'title-percentage' => 'Percentage',
-    'description-percentage' => 'Enter the percentage of the cost to allot to the resource, defaults to 100 if not provided',
-];
